requirements.txt
addict==2.4.0
ansi2html==1.8.0
attrs==23.1.0
backcall==0.2.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
einops==0.6.1
fastjsonschema==2.18.0
Flask==2.2.5
fonttools==4.38.0
fvcore==0.1.5.post20221221
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipdb==0.13.13
ipython==7.34.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_core==4.12.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
kornia==0.6.12
lpips==0.1.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.7.0
nerfacc==0.5.1
nest-asyncio==1.5.7
ninja==1.11.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
open3d==0.17.0
opencv-python==4.8.0.76
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.16.1
plyfile==0.9
portalocker==2.7.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pybind11==2.11.1
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch3d==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
rich==13.5.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yacs==0.1.8
zipp==3.15.0
